+++ lib/Sources/App/NavBarModel.swift
@@ -16,7 +16,6 @@ public final class NavBarModel {
 
     var fullscreenButtonDisabled: Bool = false
     var isSettingsPresented: Bool = false
-    /*var isInSearchMode: Bool = true*/ // need to find where we take keyboard input and flip this
 
     private(set) var isFullscreen: Bool = false
     private let tabManagementAction: () -> Void
