+++ lib/Sources/App/v2/NavBarViewV2.swift
@@ -16,14 +16,12 @@ struct NavBarViewV2: View {
     // stops when the keyboard is present vs when it's not.
     private var backgroundGradientStops: [Gradient.Stop] {
         if keyboardPresent {[
-//                Gradient.Stop(color: Color.clear, location: 0),
-            Gradient.Stop(color: Color.navigationBackground.opacity(0), location: 0),
+                Gradient.Stop(color: Color.navigationBackground.opacity(0), location: 0),
                 Gradient.Stop(color: Color.navigationBackground.opacity(0.8), location: 0.2),
                 Gradient.Stop(color: Color.navigationBackground.opacity(0.95), location: 1),
             ]
         } else {[
-//            Gradient.Stop(color: Color.clear, location: 0),
-            Gradient.Stop(color: Color.navigationBackground.opacity(0), location: 0),
+                Gradient.Stop(color: Color.navigationBackground.opacity(0), location: 0),
                 Gradient.Stop(color: Color.navigationBackground.opacity(0.95), location: 1),
             ]
         }
@@ -53,47 +51,14 @@ struct NavBarViewV2: View {
         .padding(.horizontal, 16)
         .padding(.bottom, 6)
         .frame(maxWidth: .infinity)
-//        .blur(radius: 20.0)
-
         .background(
             LinearGradient(
                 stops: backgroundGradientStops,
-                startPoint: UnitPoint(x: 0.5, y: 0),
-                endPoint: UnitPoint(x: 0.5, y: 1)
+                startPoint: .top,
+                endPoint: .bottom
             )
-//            DEAR JUSTIN. YOU ARE TRYING TO MAKE THIS VIEW TRANSPARENT BUT WITH THE GRADIENT
-//            .background(.thinMaterial)
             .edgesIgnoringSafeArea(.all)
         )
-//        .border(Color.red)
-//        .mask(
-//            LinearGradient(gradient: Gradient(colors: [Color.black, Color.black, Color.black, Color.black.opacity(0)]), startPoint: .top, endPoint: .bottom)
-//                .edgesIgnoringSafeArea(.all)
-//        )
-
-//        .background(
-//            Rectangle()
-//                // Use a material for the translucent blur effect
-//                .fill(.ultraThinMaterial)
-//                // Apply a linear gradient as a mask
-//                // The white/black areas define the visible/transparent parts of the mask
-//                .background (
-//                    LinearGradient(
-//                        stops: backgroundGradientStops,
-//                        startPoint: .top,
-//                        endPoint: .bottom
-//                    )
-//                )
-//                .mask(
-//                    LinearGradient(
-//                        stops: backgroundGradientStops,
-//                        startPoint: .top,
-//                        endPoint: .bottom
-//                    )
-//                )
-//                .ignoresSafeArea()
-//        )
-
     }
 }
