+++ lib/Sources/WebView/WebPageModel.swift
@@ -57,6 +57,7 @@ public class WebPageModel: Identifiable {
 
     func createWebView() -> WKWebView {
         let webView = WKWebView(frame: .zero, configuration: config)
+        // Ensures that content a the bottom of the webview is scrollable above the nav bar
         webView.scrollView.contentInset = UIEdgeInsets(top: 0, left: 0, bottom: 85, right: 0)
         webView.allowsBackForwardNavigationGestures = true
         return webView
